correctly send the broadcast to all packages
