+++ app/src/main/java/net/fabiszewski/ulogger/ExternalCommandReceiver.java
@@ -13,6 +13,8 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Bundle;
+import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -24,6 +26,7 @@ import net.fabiszewski.ulogger.services.LoggerService;
 import net.fabiszewski.ulogger.services.WebSyncService;
 import net.fabiszewski.ulogger.ui.AutoNamePreference;
 import net.fabiszewski.ulogger.ui.SettingsActivity;
+import net.fabiszewski.ulogger.utils.BroadcastHelper;
 
 public class ExternalCommandReceiver extends BroadcastReceiver {
 
@@ -31,6 +34,10 @@ public class ExternalCommandReceiver extends BroadcastReceiver {
     private static final String START_NEW_LOGGER = "start new logger";
     private static final String STOP_LOGGER = "stop logger";
     private static final String START_UPLOAD = "start upload";
+    private static final String GET_TRACK_ID = "get track id";
+
+    public static final String BROADCAST_TRACK_ID = "net.fabiszewski.ulogger.broadcast.track_id";
+    public static final String EXTRA_TRACK_ID = "trackId";
 
     @Override
     public void onReceive(@NonNull Context context, @Nullable Intent intent) {
@@ -48,6 +55,7 @@ public class ExternalCommandReceiver extends BroadcastReceiver {
                     case START_NEW_LOGGER -> startNewLoggerService(context, overwrite);
                     case STOP_LOGGER -> stopLogger(context);
                     case START_UPLOAD -> uploadData(context);
+                    case GET_TRACK_ID -> sendTrackId(context);
                 }
             }
         }
@@ -93,5 +101,18 @@ public class ExternalCommandReceiver extends BroadcastReceiver {
             Intent intent = new Intent(context, WebSyncService.class);
             ContextCompat.startForegroundService(context, intent);
         }
+    }
+
+    /**
+    * Send current track id via broadcast
+    * @param context Context
+    */
+    private void sendTrackId(@NonNull Context context) {
+        int trackId = DbAccess.getTrackId(context);
+        // Create intent without restricting to package (allows external receivers like Tasker)
+        Intent intent = new Intent(BROADCAST_TRACK_ID);
+        intent.putExtra(EXTRA_TRACK_ID, trackId);
+        // Send as a public broadcast (accessible to other apps)
+        context.sendBroadcast(intent);
     }
 }
